Move to internal. Rename to S3LogStoreUtil
